refactor store search failure setting and should persist check
