+++ src/main/java/com/amazon/opendistroforelasticsearch/search/asynchronous/context/active/AsynchronousSearchActiveContext.java
@@ -60,16 +60,16 @@ public class AsynchronousSearchActiveContext extends AsynchronousSearchContext i
     private final SetOnce<Exception> error;
     private final SetOnce<SearchResponse> searchResponse;
     private final AtomicBoolean closed;
+    private final Supplier<Boolean> storeSearchFailureSupplier;
     private AsynchronousSearchContextPermits asynchronousSearchContextPermits;
     private Supplier<SearchResponse> partialResponseSupplier;
     @Nullable
     private final User user;
 
     public AsynchronousSearchActiveContext(AsynchronousSearchContextId asynchronousSearchContextId, String nodeId,
-                                    TimeValue keepAlive, boolean keepOnCompletion,
-                                    ThreadPool threadPool, LongSupplier currentTimeSupplier,
-                                    AsynchronousSearchProgressListener asynchronousSearchProgressListener,
-                                    @Nullable User user) {
+                                    TimeValue keepAlive, boolean keepOnCompletion, ThreadPool threadPool, LongSupplier currentTimeSupplier,
+                                    AsynchronousSearchProgressListener asynchronousSearchProgressListener, @Nullable User user,
+                                           Supplier<Boolean> storeSearchFailureSupplier) {
         super(asynchronousSearchContextId, currentTimeSupplier);
         this.keepOnCompletion = keepOnCompletion;
         this.error = new SetOnce<>();
@@ -85,6 +85,7 @@ public class AsynchronousSearchActiveContext extends AsynchronousSearchContext i
         this.asynchronousSearchContextPermits = keepOnCompletion ? new AsynchronousSearchContextPermits(asynchronousSearchContextId,
                 threadPool) : new NoopAsynchronousSearchContextPermits(asynchronousSearchContextId);
         this.user = user;
+        this.storeSearchFailureSupplier = storeSearchFailureSupplier;
     }
 
     public void setTask(SearchTask searchTask) {
@@ -141,7 +142,7 @@ public class AsynchronousSearchActiveContext extends AsynchronousSearchContext i
     }
 
     public boolean shouldPersist() {
-        return keepOnCompletion && isExpired() == false && isAlive();
+        return keepOnCompletion && isExpired() == false && isAlive() && (error.get() == null || storeSearchFailureSupplier.get());
     }
 
     public boolean keepOnCompletion() {
