+++ src/main/java/com/amazon/opendistroforelasticsearch/search/asynchronous/service/AsynchronousSearchService.java
@@ -116,6 +116,10 @@ public class AsynchronousSearchService extends AbstractLifecycleComponent implem
             "opendistro_asynchronous_search.max_wait_for_completion_timeout", timeValueMinutes(1), Setting.Property.NodeScope,
             Setting.Property.Dynamic);
 
+    public static final Setting<Boolean> STORE_SEARCH_FAILURES_SETTING =
+            Setting.boolSetting("opendistro_asynchronous_search.store_search_failures", false,
+                    Setting.Property.NodeScope, Setting.Property.Dynamic);
+
     private volatile long maxKeepAlive;
     private volatile long maxWaitForCompletionTimeout;
     private volatile long maxSearchRunningTime;
@@ -130,6 +134,7 @@ public class AsynchronousSearchService extends AbstractLifecycleComponent implem
     private final AsynchronousSearchStateMachine asynchronousSearchStateMachine;
     private final NamedWriteableRegistry namedWriteableRegistry;
     private final AsynchronousSearchContextEventListener contextEventListener;
+    private volatile boolean storeSearchFailure;
 
     public AsynchronousSearchService(AsynchronousSearchPersistenceService asynchronousSearchPersistenceService,
                                      AsynchronousSearchActiveStore asynchronousSearchActiveStore, Client client,
@@ -146,6 +151,8 @@ public class AsynchronousSearchService extends AbstractLifecycleComponent implem
         setMaxWaitForCompletionTimeout(MAX_WAIT_FOR_COMPLETION_TIMEOUT_SETTING.get(settings));
         clusterService.getClusterSettings().addSettingsUpdateConsumer(MAX_SEARCH_RUNNING_TIME_SETTING, this::setMaxSearchRunningTime);
         setMaxSearchRunningTime(MAX_SEARCH_RUNNING_TIME_SETTING.get(settings));
+        clusterService.getClusterSettings().addSettingsUpdateConsumer(STORE_SEARCH_FAILURES_SETTING, this::setStoreSearchFailure);
+        setStoreSearchFailure(STORE_SEARCH_FAILURES_SETTING.get(clusterService.getSettings()));
         this.threadPool = threadPool;
         this.clusterService = clusterService;
         this.persistenceService = asynchronousSearchPersistenceService;
@@ -191,8 +198,8 @@ public class AsynchronousSearchService extends AbstractLifecycleComponent implem
                 threadPool::relativeTimeInMillis,
                 reduceContextBuilder);
         AsynchronousSearchActiveContext asynchronousSearchContext = new AsynchronousSearchActiveContext(asynchronousSearchContextId,
-                clusterService.localNode().getId(),
-                request.getKeepAlive(), request.getKeepOnCompletion(), threadPool, currentTimeSupplier, progressActionListener, user);
+                clusterService.localNode().getId(), request.getKeepAlive(), request.getKeepOnCompletion(), threadPool, currentTimeSupplier,
+                progressActionListener, user, () -> storeSearchFailure);
         asynchronousSearchActiveStore.putContext(asynchronousSearchContextId, asynchronousSearchContext,
                 contextEventListener::onContextRejected);
         contextEventListener.onContextInitialized(asynchronousSearchContextId);
@@ -693,5 +700,9 @@ public class AsynchronousSearchService extends AbstractLifecycleComponent implem
         } else {
             return e;
         }
+    }
+
+    private void setStoreSearchFailure(boolean storeSearchFailure) {
+        this.storeSearchFailure = storeSearchFailure;
     }
 }
