+++ src/test/java/com/amazon/opendistroforelasticsearch/search/asynchronous/restIT/AsynchronousSearchSettingsIT.java
@@ -17,7 +17,6 @@ package com.amazon.opendistroforelasticsearch.search.asynchronous.restIT;
 
 import com.amazon.opendistroforelasticsearch.search.asynchronous.context.active.AsynchronousSearchActiveStore;
 import com.amazon.opendistroforelasticsearch.search.asynchronous.context.state.AsynchronousSearchState;
-import com.amazon.opendistroforelasticsearch.search.asynchronous.processor.AsynchronousSearchPostProcessor;
 import com.amazon.opendistroforelasticsearch.search.asynchronous.request.GetAsynchronousSearchRequest;
 import com.amazon.opendistroforelasticsearch.search.asynchronous.request.SubmitAsynchronousSearchRequest;
 import com.amazon.opendistroforelasticsearch.search.asynchronous.response.AsynchronousSearchResponse;
@@ -146,7 +145,7 @@ public class AsynchronousSearchSettingsIT extends AsynchronousSearchRestTestCase
                 return e.getMessage().contains("resource_not_found");
             }
         });
-        updateClusterSettings(AsynchronousSearchPostProcessor.STORE_SEARCH_FAILURES_SETTING.getKey(),
+        updateClusterSettings(AsynchronousSearchService.STORE_SEARCH_FAILURES_SETTING.getKey(),
                 true);
         AsynchronousSearchResponse submitResponse = executeSubmitAsynchronousSearch(request);
         waitUntil(() -> {
@@ -159,7 +158,7 @@ public class AsynchronousSearchSettingsIT extends AsynchronousSearchRestTestCase
         });
         assertEquals(executeGetAsynchronousSearch(new GetAsynchronousSearchRequest(submitResponse.getId())).getState(),
                 AsynchronousSearchState.STORE_RESIDENT);
-        updateClusterSettings(AsynchronousSearchPostProcessor.STORE_SEARCH_FAILURES_SETTING.getKey(),
+        updateClusterSettings(AsynchronousSearchService.STORE_SEARCH_FAILURES_SETTING.getKey(),
                 false);
     }
 }
