+++ src/test/java/com/amazon/opendistroforelasticsearch/search/asynchronous/service/AsynchronousSearchPostProcessorTests.java
@@ -92,7 +92,7 @@ public class AsynchronousSearchPostProcessorTests extends ESTestCase {
         final Set<Setting<?>> settingsSet =
                 Stream.concat(ClusterSettings.BUILT_IN_CLUSTER_SETTINGS.stream(), Stream.of(
                         AsynchronousSearchActiveStore.MAX_RUNNING_SEARCHES_SETTING,
-                        AsynchronousSearchPostProcessor.STORE_SEARCH_FAILURES_SETTING,
+                        AsynchronousSearchService.STORE_SEARCH_FAILURES_SETTING,
                         AsynchronousSearchService.MAX_KEEP_ALIVE_SETTING,
                         AsynchronousSearchService.MAX_SEARCH_RUNNING_TIME_SETTING,
                         AsynchronousSearchService.MAX_WAIT_FOR_COMPLETION_TIMEOUT_SETTING)).collect(Collectors.toSet());
